+++ example/lib/main.dart
@@ -1,3 +1,4 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:eagle_provider/eagle_provider.dart';
 import 'package:flutter/material.dart';
 
@@ -5,104 +6,13 @@ void main() {
   runApp(const MyApp());
 }
 
-class Dependence {
-  static final homeController = HomeController();
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return ControllerProvider(
-      controllers: [Dependence.homeController],
-      child: MaterialApp(
-        title: 'Flutter Demo',
-        theme: ThemeData(
-          primarySwatch: Colors.blue,
-        ),
-        home: const MyHomePage(),
-      ),
-    );
-  }
-}
-
-class MyHomePage2 extends StatelessWidget {
-  const MyHomePage2({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    var controller = ControllerProvider.of<HomeController>(context);
-    controller.changeName('Antonio', Status.failure);
-
-    return Scaffold(
-      appBar: AppBar(),
-    );
-  }
-}
-
-class MyHomePage extends StatelessWidget {
-  const MyHomePage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    var controller = ControllerProvider.of<HomeController>(context);
-    controller.changeName('Pryscilla', Status.initial);
-
-    return Scaffold(
-      appBar: AppBar(
-        title: GestureDetector(
-          onTap: () {
-            debugPrint(controller.state.toString());
-          },
-          child: const Text('Home page'),
-        ),
-      ),
-      body: ControllerConsumer<HomeController, HomeState>(
-        // builderWhen: (before, after) {
-        //   return before.name != after.name;
-        // },
-        // child: const Center(child: CircularProgressIndicator()),
-        // listenWhen: (before, after) {},
-        listener: (before, after) {
-          debugPrint('LISTNER');
-          debugPrint(before.name);
-          debugPrint(after.name);
-        },
-        builder: (context, state) {
-          debugPrint(state.toString());
-          if (state.status == Status.loading) {
-            return const Center(child: CircularProgressIndicator());
-          }
-
-          return Center(
-            child: GestureDetector(
-              onTap: () {
-                Navigator.of(context).push(
-                  MaterialPageRoute(
-                    builder: (context) => const MyHomePage2(),
-                  ),
-                );
-              },
-              child: Text('Status ${state.status}, name: ${state.name}'),
-            ),
-          );
-        },
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          controller.changeName('firmino', Status.success);
-        },
-      ),
-    );
-  }
-}
-
-// State
 class HomeState extends StateController {
-  const HomeState({super.status, this.name});
-
   final String? name;
+
+  const HomeState({
+    super.status,
+    this.name,
+  });
 
   @override
   HomeState copyWith({
@@ -119,15 +29,105 @@ class HomeState extends StateController {
   List<Object?> get props => [status, name];
 }
 
-// Controller
 class HomeController extends Controller<HomeState> {
-  HomeController() : super(const HomeState());
+  HomeController(super.value);
 
-  changeStatus([Status? status]) {
-    emit(state.copyWith(status: status ?? Status.success));
+  @override
+  void onChange(HomeState? previous, HomeState current) {
+    print('PREVIOUS ${previous?.name}');
+    print('CURRENT ${current.name}');
   }
 
-  changeName(String name, Status status) {
-    emit(state.copyWith(name: name, status: status));
+  setName(String name) {
+    emit(value.copyWith(
+      name: name,
+      status: Status.success,
+    ));
+  }
+
+  setStatus([Status? status]) {
+    emit(value.copyWith(status: status ?? Status.failure));
+  }
+}
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return ControllerProvider(
+      controllers: [
+        HomeController(
+          const HomeState(
+            status: Status.loading,
+            name: 'Lucas',
+          ),
+        ),
+      ],
+      child: MaterialApp(
+        title: 'Flutter Demo',
+        theme: ThemeData(
+          primarySwatch: Colors.blue,
+        ),
+        home: const MyHomePage(),
+      ),
+    );
+  }
+}
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key});
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  @override
+  Widget build(BuildContext context) {
+    var controller = ControllerProvider.of<HomeController>(context);
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text(
+          'Eagle Provider',
+        ),
+        actions: [
+          ElevatedButton(
+            onPressed: () {
+              controller!.setName('Sabrina');
+              controller.setStatus(Status.loading);
+            },
+            child: const Text('change'),
+          ),
+        ],
+      ),
+      body: ControllerConsummer<HomeController, HomeState>(
+        listener: (context, state) {
+          print('LISTENER ${state.name}');
+        },
+        listenWhen: (previous, current) {
+          return previous.name != current.name;
+        },
+        buildWhen: (previous, current) {
+          return previous.name != current.name ||
+              previous.status != current.status;
+        },
+        builder: (context, state) {
+          if (state.status == Status.loading) {
+            return const Center(child: CircularProgressIndicator());
+          }
+
+          return Center(
+            child: Text(state.name.toString()),
+          );
+        },
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          // controller!.setName('Pryscilla');
+          controller!.setStatus();
+        },
+      ),
+    );
   }
 }