+++ lib/eagle_provider.dart
@@ -1,8 +1,7 @@
 library eagle_provider;
 
-export 'src/controller.dart';
-export 'src/controller.dart';
+export 'src/contracts/controller.dart';
+export 'src/contracts/state_controller.dart';
 export 'src/controller_provider.dart';
-export 'src/state_controller.dart';
 export 'src/widgets/controller_builder.dart';
 export 'src/widgets/controller_consummer.dart';